ATR-564: fixed foreign key unit tests
