+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/DacReferentialIntegrity/DacForeignKeyDeclaration/Sources/CorrectForeignKey/SOOrder.cs
@@ -3,28 +3,7 @@ using PX.Data.ReferentialIntegrity.Attributes;
 
 namespace Acuminator.Tests.Tests.StaticAnalysis.DacReferentialIntegrity.Sources
 {
-	[PXCacheName("SO Order")]
-	public partial class SOLine : PX.Data.IBqlTable
-	{
-		public class PK : PrimaryKeyOf<SOLine>.By<orderType, orderNbr, lineNbr>
-		{
-			public static SOLine Find(PXGraph graph, string orderType, string orderNbr, int? lineNbr) => FindBy(graph, orderType, orderNbr, lineNbr);
-		}
-
-		public static class FK
-		{
-			public class SOOrder : PX.Objects.SO.SOOrder.PK.ForeignKeyOf<SOLine>.By<orderType, orderNbr> { }
-		}
-
-		public abstract class orderType : IBqlField { }
-
-		public abstract class orderNbr : IBqlField { }
-
-		public abstract class lineNbr : IBqlField { }
-	}
-
-
-	[PXCacheName("SO Line")]
+	[PXHidden]
 	public class SOOrder : IBqlTable
 	{
 		public class PK : PrimaryKeyOf<SOLine>.By<orderType, orderNbr>
